+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
-
+* Fix modal table updating on input
 
 ## 1.0.0
 * Improved localization configurability